:sparkles: feat: add the create pou function
