+++ src/renderer/components/_organisms/explorer/project.tsx
@@ -1,34 +1,16 @@
-import {
-	ArrowIcon,
-	FolderIcon,
-	FunctionBlockIcon,
-	FunctionIcon,
-	PlusIcon,
-	ProgramIcon,
-} from '@root/renderer/assets'
+import { FolderIcon } from '@root/renderer/assets'
 import {
 	ProjectTreeRoot,
 	ProjectTreeBranch,
 	ProjectTreeLeaf,
 } from '@components/_molecules/project-tree'
 import { useOpenPLCStore } from '@root/renderer/store'
-import { IPouTemplate } from '@root/types/transfer'
-import {
-	Popover,
-	PopoverTrigger,
-	PopoverContent,
-	PouCard,
-	PouCardLabel,
-	PouCardForm,
-} from '../../_molecules'
+
 import _ from 'lodash'
-import { ReactElement, useState } from 'react'
 import { ITabProps } from '@root/renderer/store/slices'
 import { CreatePou } from '../../_features/[workspace]/create-pou'
 
 const Project = () => {
-	const [containerIsOpen, setContainerIsOpen] = useState(false)
-	const [createPouFormIsOpen, setCreatePouFormIsOpen] = useState(false)
 	const {
 		workspaceState: { projectData: { pous } },
 		updateEditor,
@@ -40,28 +22,6 @@ const Project = () => {
 		updateTabs(tab)
 		updateEditor({ path: tab.name, value: tab.body })
 	}
-
-	const CreatePouSources: {
-		id: number
-		label: 'function' | 'function-block' | 'program'
-		icon: ReactElement
-	}[] = [
-		{
-			id: 0,
-			label: 'function',
-			icon: <FunctionIcon size='sm' />,
-		},
-		{
-			id: 1,
-			label: 'function-block',
-			icon: <FunctionBlockIcon size='sm' />,
-		},
-		{
-			id: 2,
-			label: 'program',
-			icon: <ProgramIcon size='sm' />,
-		},
-	]
 
 	return (
 		<>
@@ -84,51 +44,6 @@ const Project = () => {
 				</div>
 				<div id='create-pou-container'>
 					<CreatePou />
-					{/* <Popover>
-						<PopoverTrigger
-							id='create-pou-trigger'
-							type='button'
-							className='w-10 h-8 rounded-lg bg-brand flex justify-center items-center'
-						>
-							<PlusIcon className='stroke-white' />
-						</PopoverTrigger>
-
-						<PopoverContent
-							alignOffset={-7}
-							sideOffset={10}
-							align='end'
-							className='w-[188px] h-fit shadow-card dark:shadow-dark-card border border-brand-light dark:border-brand-medium-dark bg-white dark:bg-neutral-950 p-2 rounded-lg flex flex-col gap-2'
-						>
-							{CreatePouSources.map(({ id, label, icon }) => (
-								<Popover key={id}>
-									<PopoverTrigger id={`create-${label}-trigger`}>
-										<div
-											id={`create-${label}-trigger-container`}
-											className='data-[state=open]:bg-neutral-100 py-[2px] px-[6px] justify-between dark:data-[state=open]:bg-neutral-900 relative flex items-center w-full h-7 gap-[6px] rounded-md cursor-pointer select-none hover:bg-neutral-100 dark:hover:bg-neutral-900'
-										>
-											{icon}
-											<p className='text-start font-caption text-xs font-normal text-neutral-1000 dark:text-neutral-300 flex-1 my-[2px]'>
-												{_.startCase(label)}
-											</p>
-											<ArrowIcon size='md' direction='right' />
-										</div>
-									</PopoverTrigger>
-									<PopoverContent
-										id={`create-${label}-content`}
-										sideOffset={14}
-										alignOffset={-7}
-										align='start'
-										side='right'
-									>
-										<PouCard>
-											<PouCardLabel type={label} />
-											<PouCardForm type={label} />
-										</PouCard>
-									</PopoverContent>
-								</Popover>
-							))}
-						</PopoverContent>
-					</Popover> */}
 				</div>
 			</div>
 			{/* Data display */}
